# [tests] Speed up TestWizardModel